requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
python-dotenv==1.0.1
requests==2.32.5
sqlalchemy
psycopg2-binary
apify-client
